requirements (3).txt
torch
torchvision
matplotlib
numpy
pillow
opencv-python
requests
gradio